+++ CHANGELOG.md
@@ -2,3 +2,8 @@
 
 ## 1.0.0.1
 - Initial release
+
+## 1.0.0.2
+- Added support for filtering logs (Hide TRACE, Hide DEBUG)
+- Clearing filters when loading new file
+- minor UI improvements